fix: handle and wrap stripe fetches in neverthrow
