+++ src/app/modules/payments/stripe.controller.ts
@@ -14,7 +14,6 @@ import {
   ReconciliationEventsReportLine,
   ReconciliationReport,
 } from '../../../../shared/types'
-import { IPopulatedForm } from '../../../types'
 import config from '../../config/config'
 import { paymentConfig } from '../../config/features/payment.config'
 import { createLoggerWithLabel } from '../../config/logger'
@@ -33,7 +32,7 @@ import {
   StripeMetadataIncorrectEnvError,
 } from './stripe.errors'
 import * as StripeService from './stripe.service'
-import { convertToInvoiceFormat, mapRouteError } from './stripe.utils'
+import { mapRouteError } from './stripe.utils'
 
 const logger = createLoggerWithLabel(module)
 
@@ -191,7 +190,7 @@ export const downloadPaymentInvoice: ControllerHandler<{
     PaymentService.findPaymentById(paymentId),
     FormService.retrieveFullFormById(formId).andThen(checkFormIsEncryptMode),
   ])
-    .map(([payment, populatedForm]) => {
+    .andThen(([payment, populatedForm]) => {
       logger.info({
         message: 'Found paymentId in payment document',
         meta: {
@@ -199,63 +198,14 @@ export const downloadPaymentInvoice: ControllerHandler<{
           payment,
         },
       })
-      if (!payment.completedPayment?.receiptUrl) {
-        return res
-          .status(StatusCodes.NOT_FOUND)
-          .send({ message: 'Receipt url not ready' })
-      }
-      // retrieve receiptURL as html
-      return (
-        axios
-          .get<string>(payment.completedPayment.receiptUrl)
-          // convert to pdf and return
-          .then((receiptUrlResponse) => {
-            const html = receiptUrlResponse.data
-            const agencyBusinessInfo = (populatedForm as IPopulatedForm).admin
-              .agency.business
-            const formBusinessInfo = populatedForm.business
-
-            const businessAddress = [
-              formBusinessInfo?.address,
-              agencyBusinessInfo?.address,
-            ].find(Boolean)
-
-            const businessGstRegNo = [
-              formBusinessInfo?.gstRegNo,
-              agencyBusinessInfo?.gstRegNo,
-            ].find(Boolean)
-
-            // we will still continute the invoice generation even if there's no address/gstregno
-            if (!businessAddress || !businessGstRegNo)
-              logger.warn({
-                message:
-                  'Some business info not available during invoice generation. Expecting either agency or form to have business info',
-                meta: {
-                  action: 'downloadPaymentInvoice',
-                  payment,
-                  agencyName: populatedForm.admin.agency.fullName,
-                  agencyBusinessInfo,
-                  formBusinessInfo,
-                },
-              })
-            const invoiceHtml = convertToInvoiceFormat(html, {
-              address: businessAddress || '',
-              gstRegNo: businessGstRegNo || '',
-              formTitle: populatedForm.title,
-              submissionId: payment.completedPayment?.submissionId || '',
-            })
-
-            const pdfBufferPromise = generatePdfFromHtml(invoiceHtml)
-            return pdfBufferPromise
-          })
-          .then((pdfBuffer) => {
-            res.set({
-              'Content-Type': 'application/pdf',
-              'Content-Disposition': `attachment; filename=${paymentId}-invoice.pdf`,
-            })
-            return res.status(StatusCodes.OK).send(pdfBuffer)
-          })
-      )
+      return StripeService.generatePaymentInvoice(payment, populatedForm)
+    })
+    .map((pdfBuffer) => {
+      res.set({
+        'Content-Type': 'application/pdf',
+        'Content-Disposition': `attachment; filename=${paymentId}-invoice.pdf`,
+      })
+      return res.status(StatusCodes.OK).send(pdfBuffer)
     })
     .mapErr((error) => {
       logger.error({
